Conflicts:
	publications/radical_rutgers.bib
	publications/radical_rutgers.bib~

:x

+++ misc/redishes/redishes.pl
@@ -5,9 +5,61 @@ BEGIN {
   use Net::Daemon;
 }
 
+my $HOST = `hostname -f`;
+chomp ($HOST);
+
+################################################################################
+#
+my $help = <<EOT;
+
+  This service manages on-demand redis service instances.  Instances can be
+  created, listed, and killed (canceled).  Creation requires a password.
+  Instances are shut down autimatically after some timeout (default: 1 day).  
+
+  Commands:
+  ---------
+
+    HELP
+          This message.
+
+    CREATE SECRET=<secret>
+           [PASS  = ""          ]
+           [TTL   = 60 * 60 * 24]
+           [PORT  = <auto>      ]
+          Creates a new redis instance on <PORT> with <PASS> and for <TTL> seconds.
+          <SECRET> is a password required to create a new redis instance.
+
+    EXTENT [PASS=<pass>] id
+          Restart ttl counter for server <id>.
+
+    CANCEL [PASS=<pass>] id
+          Kill server <id>.
+
+    PURGE [PASS=<pass>] id
+          Remove state of server <id>.
+
+    PURGE SECRET=<secret>
+          Remove state of all finished servers.
+
+    STATUS
+          List status for all servers.
+
+    STATUS id
+          Show status for server <id>.
+
+    SHUTDOWN SECRET=<secret>
+          Kill this very service.
+
+  Example:
+  --------
+
+    telnet 
+
+EOT
 
 
-######################################################################
+
+################################################################################
 #
 # the server module
 #
@@ -22,12 +74,14 @@ our @ISA = "Net::Daemon";
 use Data::Dumper;
 use POSIX ":sys_wait_h";
 
+#---------------------------------------------------------------------
+$ENV{'PATH'} = "/bin/:/usr/bin/:/usr/local/bin/";
 
 my $SERVER_BIN   = 'redis-server';
-my $SERVER_LIMIT = 128;
 my $SERVER_TTL   = 60 * 60 * 24;  # one day
+my $SERVER_LIMIT = 128;
 my $PORT_MIN     = 10000;
-my $PORT_MAX     = $PORT_MIN + $SERVER_LIMIT;
+my $PORT_MAX     = $PORT_MIN + $SERVER_LIMIT + 1;
 my $ROOT         = "/tmp/redishes/";
 my $SECRET       = $ENV{'REDISHES_SECRET'} or die "please set 'REDISHES_SECRET' before running\n";
 
@@ -47,11 +101,18 @@ sub new ($$;$)
 
   bless ($self, $type);
 
-  if ( ! fork () )
+  my $tmp_pid = fork ();
+  if ( ! $tmp_pid )
   {
-    $self->cleaner ();
+    if ( ! fork () )
+    {
+      $self->cleaner ();
+      exit (0)
+    }
+    exit (0);
   }
 
+  waitpid ($tmp_pid, 0);
 
   return $self;
 }
@@ -60,80 +121,68 @@ sub cleaner ($)
 {
   while ( 1 )
   {
-    my $purged = 0;
     my $active = 0;
 
     # take care not to purge 'ports/'
     my @server = glob ("$ROOT/*-*/");
 
-    PURGE:
+    CANCEL:
     foreach my $pwd ( @server )
     {
-      if ( -e "$pwd/purged" )
-      {
-        next PURGE;
-      }
+      next CANCEL if -e "$pwd/canceled";
+      next CANCEL if -e "$pwd/timeout";
+      next CANCEL if -e "$pwd/done";
+      next CANCEL if -e "$pwd/killed";
+      next CANCEL if -e "$pwd/timeout";
 
-      my $purge = 0;
+      my $cancel = 0;
 
-      if ( -e "$ROOT/action.shutdown" )
-      {
-        $purge = 1;
-      }
-      elsif ( -e "$pwd/action.purge" )
-      {
-        $purge = 1;
-      }
-      elsif ( not -e "$pwd/redis.ttl" )
-      {
-        $purge = 1;
-      }
+      if    (     -e "$ROOT/action.shutdown" ) { $cancel = 1; }
+      elsif (     -e "$pwd/action.cancel"    ) { $cancel = 1; }
+      elsif ( not -e "$pwd/ttl"              ) { $cancel = 2; }
       else
       {
-        my $ttl = `cat $pwd/redis.ttl`;  chomp ($ttl);
-
-        my $ctime = ( stat "$pwd/redis.pid" )[10] || next PURGE;
+        my $ttl   = `cat $pwd/ttl`;  chomp ($ttl);
+        my $ctime = ( stat "$pwd/pid" )[10] || next CANCEL;
         my $now   = time;
 
-        if ( $now - $ctime > $ttl )
-        {
-          $purge = 1;
-        }
+        if ( $now - $ctime > $ttl )            { $cancel = 2; }
       }
 
-      if ( $purge )
+
+      if ( $cancel )
       {
         my $pid  = undef;
         my $port = undef;
 
-        if ( -e  "$pwd/redis.pid"  ) { $pid  = `cat $pwd/redis.pid`;   chomp ($pid);  }
-        if ( -e  "$pwd/redis.port" ) { $port = `cat $pwd/redis.port`;  chomp ($port); }
+        if ( -e  "$pwd/pid"  ) { $pid  = `cat $pwd/pid`;   chomp ($pid);  }
+        if ( -e  "$pwd/port" ) { $port = `cat $pwd/port`;  chomp ($port); }
 
         # print "pid: $pid $pwd\n";
 
         if ( defined $pid )
         {
-          kill  (2, $pid); # INT
-          sleep (1);
-          kill  (9, $pid); # KILL
+          kill  ( 9, $pid ); # KILL
+          kill  (15, $pid ); # TERM
         }
-
+          
         if ( defined $port )
         {
-          `rm    $ROOT/ports/$port`;
+          `rm $ROOT/ports/$port`;
         }
 
-        `touch $pwd/purged`;
-        `rm -f $pwd/action.purge`;
+        `rm -f $pwd/action.* >/dev/null 2>&1`;   
 
-        if ( -e "$ROOT/action.shutdown" )
+        if ( $cancel == 1 )
         {
-          `rm -rf $pwd`;
+          print " - canceled $pwd\n";
+          `touch $pwd/canceled`;   
         }
-
-        $purged ++;
-
-        print " - purged $pwd : $pid / $port\n";
+        elsif ( $cancel == 2 )
+        {
+          print " - timeout $pwd\n";
+          `touch $pwd/timeout`;   
+        }
       }
       else
       {
@@ -143,12 +192,28 @@ sub cleaner ($)
 
     if ( -e "$ROOT/action.shutdown" && $active == 0 )
     {
-      `rm -rf $ROOT`;
-      print " - shutdown $ROOT\n";
+      if ( $active )
+      {
+        print " - shutdown delayed -- $active active services found\n";
+      }
+      else
+      {
+        `rm -rf $ROOT`;
+        print " - shutdown $ROOT\n";
+        exit (0);
+      }
+    }
+
+    if ( -e "$ROOT/action.quit" )
+    {
+      sleep (2);
+      `rm -rf $ROOT/action.quit`;
+      print " - quit $ROOT\n";
       exit (0);
     }
 
     sleep (1);
+    # print "cleaning $ROOT/action.quit\n";
   }
 }
 
@@ -157,19 +222,97 @@ sub cleaner ($)
 # each client connection.
 sub Run ($) 
 {
-  my $self   = shift;
+  my $self = shift;
 
   # read from client socket
   my $sock = $self->{'socket'};
   
-  while ( defined (my $line = $sock->getline ()) ) 
+
+  $sock->print ("SECRET > "); 
+  $sock->flush ();
+
+  SECRET_LINE:
+  while ( defined (my $tainted_line = $sock->getline ()) ) 
   {
-    my $ret = undef;
+    # Remove CRLF
+    $tainted_line =~ s/\r\n//iog;
+
+    # remove taint
+    my $line = "";
+
+    if ( $tainted_line =~ /^\s*(
+                            (?:[\s0-9a-z\-_=\+\(\)\{\}\[\];:\<\>\~]+)?
+                            )\s*
+                           $/ixo)
+    {
+      $line = $1;
+      print "--$line--\n";
+
+      if ( $line eq $SECRET )
+      {
+        last SECRET_LINE;
+        print " - correct secret: '$line'\n";
+        $sock->print (" > 202 authorized\n"); 
+      }
+      else
+      {
+        # starve eventual passwd space scans
+        sleep (1);
+        print " - incorrect secret: '$line'\n";
+        $sock->print (" > 401 not authorized\n"); 
+        $sock->print ("SECRET > "); 
+        $sock->flush ();
+      }
+    }
+    else
+    {
+      # starve eventual passwd space scans
+      sleep (1);
+      print " - parse error: '" . $tainted_line . "'\n";
+      $sock->print (" > 406 parse error: '" . $tainted_line . "'\n");
+      $sock->print ("SECRET > "); 
+      $sock->flush ();
+    }
+  }
+
+  $sock->print ("REDIS  > "); 
+  $sock->flush ();
+
+  LINE:
+  while ( defined (my $tainted_line = $sock->getline ()) ) 
+  {
+    chomp ($tainted_line); # Remove CRLF 
+
+    # remove taint
+    my $line = "";
+
+    if ( $tainted_line =~ /^\s*(
+                            (?:HELP|CREATE|EXTENT|CANCEL|PURGE|STATUS|SHUTDOWN|QUIT)
+                            (?:\s+[0-9a-z\-_=\+\(\)\{\}\[\];:\<\>\~]+)?
+                            )\s*
+                           $/ixo)
+    {
+      $line = $1;
+    }
+    else
+    {
+      print " - parse error: '" . $tainted_line . "'\n";
+      $sock->print (" > 406 parse error: '" . $tainted_line . "'\n");
+      $sock->print ("REDIS  > "); 
+      $sock->flush ();
+      next LINE;
+    }
+
+
+    my $ret = "";
     my $ttl = $SERVER_TTL;
 
-    chomp $line; # Remove CRLF
+    if ( $line =~ /^\s*HELP\s*$/io )
+    {
+      $ret = $help;
+    }
 
-    if ( $line =~ /^\s*REDIS\s+CREATE(?:\s+(\S.*?))?\s*$/io )
+    elsif ( $line =~ /^\s*CREATE\s*(.*?)?\s*$/io )
     {
       my $opts = $1 || "";
       my $key  = `uuidgen`;
@@ -178,7 +321,7 @@ sub Run ($)
       my @server = glob ("$ROOT/*-*/");
 
       if ( scalar (@server) >= $SERVER_LIMIT ) {
-        $ret = "429 insufficient resources for new redis instance, try again later.";
+        $ret = "429 insufficient resources for new redis instance, try again later.\n";
       }
 
       else {
@@ -187,96 +330,169 @@ sub Run ($)
     }
 
 
-    elsif ( $line =~ /^\s*REDIS\s+EXTENT\s+(\S+)\s*$/io )
+    elsif ( $line =~ /^\s*EXTENT\s+(\S+)\s*$/io )
     {
       my $key = $1;
 
       if ( ! -d "$ROOT/$key" )
       {
-        $ret = "404 redis instance '$key' not found.";
+        $ret = "404 redis instance '$key' not found.\n";
       }
       else
       {
-        $ret = "200 redis instance '$key' revitalized.";
-        `touch $ROOT/$key/redis.pid`;
+        my $pwd = "$ROOT/$key";
+
+        if ( -e "$pwd/killed"   or
+             -e "$pwd/canceled" or
+             -e "$pwd/timeout"  or
+             -e "$pwd/done"     )
+        {
+          $ret = "409 redis instance '$key' is pining for the fjords.\n";
+        }
+        else
+        {
+          $ret = "200 redis instance '$key' revitalized.\n";
+          `touch $ROOT/$key/pid`;
+        }
       }
     }
 
 
-    elsif ( $line =~ /^\s*REDIS\s+PURGE\s+(\S+)\s*$/io )
+    elsif ( $line =~ /^\s*CANCEL\s+(\S+)\s*$/io )
     {
-      my $key = $1;
+      my $key    = $1;
 
       if ( ! -d "$ROOT/$key" ) 
       {
-        $ret = "404 redis instance '$key' not found.";
+        $ret = "404 redis instance '$key' not found.\n";
       }
-
       else 
       {
-        $ret = "202 redis instance '$key' will be purged.";
-        `touch $ROOT/$key/action.purge`;
+        $ret = "202 redis instance '$key' will be canceled.\n";
+        `touch $ROOT/$key/action.cancel`;
       }
     }
 
     
-    elsif ( $line =~ /^\s*REDIS\s+LIST\s*$/io )
+    elsif ( $line =~ /^\s*PURGE\s+(\S+)\s*$/io )
     {
-      my @server = split (/\s+/, `cd $ROOT && ls -d *-*`);
-      foreach my $server ( @server )
+      my $key    = $1;
+      my $pwd    = "$ROOT/$key";
+
+      if ( ! -d $pwd ) 
       {
-        $ret .= " < $server - ";
-        $ret .= `cat $ROOT/$server/redis.url`;
+        $ret = "404 redis instance '$key' not found.\n";
+      }
+
+      else 
+      {
+        if ( -e "$pwd/killed"   or
+             -e "$pwd/canceled" or
+             -e "$pwd/timeout"  or
+             -e "$pwd/done"     )
+        {
+          `test -e "$pwd/port && rm -f rm $ROOT/ports/\`cat $pwd/port\``;
+          `rm -rf $pwd`;
+          $ret = "200 redis instance '$key' purged.\n";
+        }
+        else 
+        {
+          $ret = "409 redis instance '$key' still running.\n";
+        }
       }
     }
 
+    
+    elsif ( $line =~ /^\s*PURGE\s*$/io )
+    {
+      my @pwds = glob ("$ROOT/*-*/");
+      for my $pwd ( @pwds )
+      {
+        if ( -e "$pwd/killed"   or
+             -e "$pwd/canceled" or
+             -e "$pwd/timeout"  or
+             -e "$pwd/done"     )
+        {
+          `test -e "$pwd/port && rm -f rm $ROOT/ports/\`cat $pwd/port\``;
+          `rm -rf $pwd`;
+          $ret .= "200 redis instance '$pwd' purged.\n";
+        }
+      }
+    }
 
-    elsif ( $line =~ /^\s*REDIS\s+STATUS\s*$/io )
+    
+    elsif ( $line =~ /^\s*STATUS\s*$/io )
     {
       my @keys = `cd $ROOT && ls -d *-*`;
       foreach my $key ( @keys )
       {
         chomp ($key);
 
-        my $status = "running";
+        my $status = "unknown";
 
-        if ( -e "$ROOT/$key/purged" ) { $status = "purged"; }
+        if ( -e "$ROOT/$key/running" ) { $status = "running";  }
+        if ( -e "$ROOT/$key/killed"  ) { $status = "killed";   }
+        if ( -e "$ROOT/$key/done"    ) { $status = "done";     }
+        if ( -e "$ROOT/$key/canceled") { $status = "canceled"; }
+        if ( -e "$ROOT/$key/timeout" ) { $status = "timeout";  }
 
-        my $pid  = `cat $ROOT/$key/redis.pid`;  chomp ($pid);
-        my $ttl  = `cat $ROOT/$key/redis.ttl`;  chomp ($ttl);
-        my $port = `cat $ROOT/$key/redis.port`; chomp ($port);
-        my $url  = `cat $ROOT/$key/redis.url`;  chomp ($url);
+        my $pid  = `cat $ROOT/$key/pid`;  chomp ($pid);
+        my $ttl  = `cat $ROOT/$key/ttl`;  chomp ($ttl);
+        my $port = `cat $ROOT/$key/port`; chomp ($port);
+        my $url  = `cat $ROOT/$key/url`;  chomp ($url);
 
-        $ret .= sprintf (" < %s : %6d : %6d : %-8s : %s\n", 
+        $ret .= sprintf ("%s : %6s : %6s : %-8s : %s\n", 
                          $key, $pid, $port, $status, $url);
       }
     }
 
 
-    elsif ( $line =~ /^\s*REDIS\s+STATUS\s+(\S+)\s*$/io )
+    elsif ( $line =~ /^\s*STATUS\s+(\S+)\s*$/io )
     {
       my $key = $1;
 
       if ( ! -d "$ROOT/$key" )
       {
-        $ret = "404 redis instance '$key' not found.";
+        $ret = "404 redis instance '$key' not found.\n";
       }
       else
       {
-        my $pid = `cat $ROOT/$key/redis.pid`;  chomp ($pid);
-        print "pid: 'pid'\n";
-        $ret = `ls -la $ROOT/$key`;
-        $ret = `ps -lf -p $pid`;
-        print $ret;
+        my $status = "unknown";
+
+        if ( -e "$ROOT/$key/running" ) { $status = "running";  }
+        if ( -e "$ROOT/$key/killed"  ) { $status = "killed";   }
+        if ( -e "$ROOT/$key/done"    ) { $status = "done";     }
+        if ( -e "$ROOT/$key/canceled") { $status = "canceled"; }
+        if ( -e "$ROOT/$key/timeout" ) { $status = "timeout";  }
+
+        my $pid  = `cat $ROOT/$key/pid`;  chomp ($pid);
+        my $ttl  = `cat $ROOT/$key/ttl`;  chomp ($ttl);
+        my $port = `cat $ROOT/$key/port`; chomp ($port);
+        my $url  = `cat $ROOT/$key/url`;  chomp ($url);
+
+        $ret .= sprintf ("%s : %6s : %6s : %-8s : %s\n", 
+                         $key, $pid, $port, $status, $url);
       }
     }
 
-    elsif ( $line =~ /^\s*REDIS\s+SHUTDOWN\s*$/io )
+    elsif ( $line =~ /^\s*SHUTDOWN\s*$/io )
     {
       `touch $ROOT/action.shutdown`;
-      $ret = "202 service will shut down";
-      print $sock "$ret\n";
+      $ret = "202 service will shut down\n";
+      $sock->print ("$ret\n");
+
+      print "$ret\n";
+      exit  (0);
+    }
+
+    elsif ( $line =~ /^\s*QUIT\s*$/io )
+    {
+      $ret = "202 bye\n";
+      $sock->print ("$ret\n");
+      $sock->flush ();
+      `killall redishes.pl`;
       sleep (1);
+      `killall -9 redishes.pl`;
       exit  (0);
     }
 
@@ -285,12 +501,13 @@ sub Run ($)
       $ret = "418 I'm a teapot.";
     }
 
-    $ret =~ s/^ <\s*//o ;
+    my $log = $ret;
+    $log =~ s/^/ > /omg ;
 
     print " > $line\n";
-    print " < $ret\n";
+    print "$log\n";
 
-    my $rc = print $sock "$ret\n";
+    my $rc = $sock->print ("$ret\n");
     if ( ! $rc )
     {
       print "error: " . $sock->error () . "\n";
@@ -298,6 +515,9 @@ sub Run ($)
       $sock->close ();
       return;
     }
+
+    $sock->print ("REDIS  > "); 
+    $sock->flush ();
   }
 
   if ( $sock->error () ) 
@@ -322,31 +542,22 @@ sub run_server ($$$)
   my $ttl      = $SERVER_TTL;
   my $port     = -1;
   my $pwd      = "$ROOT/$key";
-  my $db       = "$pwd/redis.db";
-  my $log      = "$pwd/redis.log";
-  my $conf     = "$pwd/redis.conf";
+  my $db       = "$pwd/db";
+  my $log      = "$pwd/log";
+  my $conf     = "$pwd/conf";
   my $pass     = "";
   my $confpass = "";
-  my $secret   = "";
 
   mkdir ($pwd) or die "Cannot create dir: $!\n";
 
   # print "opts: $opts\n";
   # print "pwd : $pwd\n";
 
-  if ( $opts  =~ /\bSECRET\s*=\s*(\S+)\b/io ) { $secret = $1; }
   if ( $opts  =~ /\bTTL\s*=\s*(\d+)\b/io )    { $ttl    = $1; }
   if ( $opts  =~ /\bPASS\s*=\s*(\S+)\b/io )   { $pass   = $1; }
   if ( $opts  =~ /\bPORT\s*=\s*(\d+)\b/io )   { $port   = $1; }
 
-  # print "port: $port ($secret - $ttl - $pass)\n";
-
-  if ( ! defined ($secret) or $secret ne $SECRET )
-  {
-    # starve eventual passwd space scans
-    sleep (1);
-    return "401 Unauthorized: invalid secret provided";
-  }
+  # print "port: $port ($ttl - $pass)\n";
 
   if ( $port == -1 ) { $port = $self->find_port ($key); }
   if ( $pass       ) { $confpass = "requirepass $pass"; }
@@ -354,7 +565,7 @@ sub run_server ($$$)
 
   open (CONF, ">$conf") or die "cannot write config file '$conf': $!\n";
   print CONF <<EOT;
-daemonize yes
+daemonize no
 port $port
 timeout $ttl
 loglevel notice
@@ -370,7 +581,6 @@ maxclients 128
 appendonly no
 appendfsync everysec
 vm-enabled no
-glueoutputbuf yes
 hash-max-zipmap-entries 64
 hash-max-zipmap-value 512
 activerehashing yes
@@ -378,11 +588,45 @@ $confpass
 EOT
   close (CONF);
 
-  `echo $port > $pwd/redis.port`;
-  `echo $ttl  > $pwd/redis.ttl`;
+  `echo $port > $pwd/port`;
+  `echo $ttl  > $pwd/ttl`;
+  `touch        $pwd/running`;
 
-  # new process, start redis server
-  system ("$SERVER_BIN $conf");
+  # double fork to avoid zombies
+  my $tmp_pid = fork ();
+  if ( ! $tmp_pid )
+  {
+    if ( ! fork () )
+    {
+      my $retval = system ("$SERVER_BIN $conf");
+
+      if ( $retval == -1 ) 
+      {
+        `echo "failed to execute: $!" >> $pwd/log`;
+      }
+      elsif ( $retval & 127 )
+      {
+        my $msg = sprintf ("child died with signal %d", ($retval & 127));
+        `echo "$msg" >> $pwd/log`;
+        `touch $pwd/killed`;
+      }
+      else 
+      {
+        my $exitval = $retval >> 8;
+        my $msg = sprintf ("child exited with value %d", $exitval);
+
+        `echo "$msg" >> $pwd/log`;
+
+        if ( $exitval == 0 ) { `touch $pwd/done`  ; }
+        if ( $exitval != 0 ) { `touch $pwd/failed`; }
+      }
+
+      exit (0);
+    }
+    exit (0);
+  }
+
+  waitpid ($tmp_pid, 0);
 
   do 
   {
@@ -391,25 +635,25 @@ EOT
 
   # store pid for convenience
   my $pid = `ps -e -o pid,cmd| grep $key | grep -v grep | cut -c 1-6`; chomp ($pid);
-  `echo $pid  > $pwd/redis.pid`;
+  `echo $pid  > $pwd/pid`;
 
   my $url = "";
   my $ret  = "201 creating redis instance '$key': ";
   
   if ( $pass ) 
   { 
-    $url = "redis://:XXXXX\@localhost:$port/"; 
-    $ret = "redis://:$pass\@localhost:$port/"; 
+    $url = "redis://:XXXXX\@$HOST:$port/"; 
+    $ret = "redis://:$pass\@$HOST:$port/"; 
   }
   else         
   {
-    $url = "redis://localhost:$port/"; 
-    $ret = "redis://localhost:$port/"; 
+    $url = "redis://$HOST:$port/"; 
+    $ret = "redis://$HOST:$port/"; 
   }
 
-  `echo $url  > $pwd/redis.url`;
+  `echo $url  > $pwd/url`;
 
-  return $ret;
+  return "$key - $ret\n";
 }
 
 
@@ -418,7 +662,7 @@ sub find_port ($$)
   my $self  = shift;
   my $key   = shift;
 
-  foreach my $p ( $PORT_MIN..$PORT_MAX )
+  foreach my $p ( ($PORT_MIN+1)..$PORT_MAX )
   {
     if ( ! -e "$ROOT/ports/$p" )
     {
@@ -434,6 +678,6 @@ sub find_port ($$)
 package main;
 
 Redishes->new ({'pidfile'   => 'none',
-                'localport' => 2000,
+                'localport' => $PORT_MIN,
                 'mode'      => 'single'})-> Bind ();
